% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3PlantUML.R
\name{as.plantuml}
\alias{as.plantuml}
\title{Convert data, if possible, to a suitable PlantUML S3 Object}
\usage{
as.plantuml(uml)
}
\arguments{
\item{uml}{Data convertible to an String with a plantuml definition}
}
\value{
An PlantUML S3 object
}
\description{
Convert data, if possible, to a suitable PlantUML S3 Object
}
\details{
Throw an plantUMLErr expection
}
