% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localDev.R
\docType{data}
\name{pkgEnv}
\alias{pkgEnv}
\title{Mark the project as in local development}
\format{
An object of class \code{environment} of length 1.
}
\usage{
pkgEnv
}
\description{
Algunos tests no se pueden realizar en los sistemas de integracion continua
por que no tenemos la seguridad de tener el entorno correcto:
JVM
Graphviz
Etc
}
\details{
Este fichero debe tener la variable pkgEnv$local a FALSE en github
Para ello se recomienda subirlo la primera vez y ponerlo en .gitignore
}
\keyword{datasets}
