% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantumlr.R
\docType{class}
\name{PLANTUML}
\alias{PLANTUML}
\title{An example R6 class}
\format{
An \code{R6Class} generator object
}
\arguments{
\item{data}{una definicion del diagrama o un fichero con la misma}

\item{force}{Ver  \code{vignette("Caching diagrams")}}

\item{caption}{El texto del link}
}
\description{
plot the diagram

Genera un link a la imagen en fichero
}
\details{
\preformatted{La funcion identifica el tipo de informcion pasada en base a la
siguiente logica:
}

\enumerate{
\item Si data empieza por "/" o "x:" se considera un fichero absoluto
\item Si data contiene espacios o retornos de carro se considera dato
\item Si no se considera un fichero
}
}
\section{Methods}{

\describe{
\item{\code{initialize(...)}}{
Crea un objeto PLANTUML
Acepta un conjunto de parametros nombre=valor_cadena donde nombre
identifica un parametro de configuracion
}
}
An example R6 class
}

\examples{
\dontrun {
   plant = PLANTUML$new()
   data="@startuml
               [*]-->STATE1
             @enduml
            "
   plant$plot(data)
   plant$plot("fileName")
   plant$plot("/home/absolute_file_name")
}
}
\keyword{"plantuml"}
