% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3PlantUML.R
\name{plot.S3PlantUML}
\alias{plot.S3PlantUML}
\title{plot a diagram definition using PlantUML package}
\usage{
\method{plot}{S3PlantUML}(uml)
}
\arguments{
\item{uml}{A file or string containing the UML definition
attribute class must set to PlantUML}
}
\description{
plot a diagram definition using PlantUML package
}
\seealso{
Other generics: 
\code{\link{PLANTUML}}
}
\concept{generics}
