% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3PlantUML.R
\name{readPlantUML}
\alias{readPlantUML}
\title{Read a text file containing an UML definition according PlantUML tool}
\usage{
readPlantUML(fileName = NULL)
}
\arguments{
\item{fileName}{The file to read}
}
\value{
A PlantUML object
}
\description{
Read a text file containing an UML definition according PlantUML tool
}
